# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Up2You API Documentation",
        "description": "Documentazione delle API per la gestione di eventi e iscritti",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/events/{event_id}/register": {
            "post": {
                "tags": [
                    "Participation"
                ],
                "summary": "Registra un iscritto a un evento",
                "operationId": "11637fa8dbbbccc0ce40a360ad5864ee",
                "parameters": [
                    {
                        "name": "event_id",
                        "in": "path",
                        "description": "ID dell'evento",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "attendee_id"
                                ],
                                "properties": {
                                    "attendee_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Registrazione completata con successo"
                    },
                    "400": {
                        "description": "Errore: limite di partecipanti raggiunto o iscritto già registrato",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Numero massimo di partecipanti raggiunto"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "api_key": {
                "type": "apiKey",
                "description": "API Key necessaria per accedere alle API",
                "name": "X-API-KEY",
                "in": "header"
            }
        }
    },
...
}